requirements.txt
streamlit
pandas
aiohttp
requests
dnspython

